requirements.txt
torch
transformers
opencv-python
pillow
python-dotenv
ultralytics
onnx
onnxruntime
supervision
onnxruntime-gpu
numpy